Renamed directory and removed envelope from API call
