+++ packages/api-core/src/me-sites-plugins/fetchers.ts
@@ -2,8 +2,5 @@ import { wpcom } from '../wpcom-fetcher';
 import type { PluginsResponse } from './types';
 
 export async function fetchUserSitesPlugins(): Promise< PluginsResponse > {
-	return await wpcom.req.get(
-		{ path: '/me/sites/plugins', apiVersion: '1.1' },
-		{ http_envelope: 1 }
-	);
+	return await wpcom.req.get( { path: '/me/sites/plugins', apiVersion: '1.1' } );
 }
